biggest update: Kernel for ARM

+++ release/src-rt-6.x.4708/linux/linux-2.6.36/arch/m68k/mac/macboing.c
@@ -150,19 +150,10 @@ static void mac_init_asc( void )
 	mac_asc_inited = 1;
 }
 
-/*
- * Called to make noise; current single entry to the boing driver.
- * Does the job for simple ASC, calls other routines else.
- * XXX Fixme:
- * Should be split into asc_mksound, easc_mksound, av_mksound and
- * function pointer set in mac_init_asc which would be called at
- * init time.
- * _This_ is rather ugly ...
- */
 void mac_mksound( unsigned int freq, unsigned int length )
 {
 	__u32 cfreq = ( freq << 5 ) / 468;
-	__u32 flags;
+	unsigned long flags;
 	int i;
 
 	if ( mac_special_bell == NULL )
@@ -224,7 +215,7 @@ static void mac_nosound( unsigned long ignored )
  */
 static void mac_quadra_start_bell( unsigned int freq, unsigned int length, unsigned int volume )
 {
-	__u32 flags;
+	unsigned long flags;
 
 	/* if the bell is already ringing, ring longer */
 	if ( mac_bell_duration > 0 )
@@ -271,7 +262,7 @@ static void mac_quadra_start_bell( unsigned int freq, unsigned int length, unsig
 static void mac_quadra_ring_bell( unsigned long ignored )
 {
 	int	i, count = mac_asc_samplespersec / HZ;
-	__u32 flags;
+	unsigned long flags;
 
 	/*
 	 * we neither want a sound buffer overflow nor underflow, so we need to match
