+++ release/src-rt-6.x.4708/linux/linux-2.6.36/include/linux/netfilter_ipv4/ipt_time.h
@@ -3,12 +3,12 @@
 
 
 struct ipt_time_info {
-	u_int8_t days_match;	/* 1 bit per day (bit 0 = Sunday) */
-	u_int32_t time_start;	/* 0 < time_start < 24*60*60-1 = 86399 */
-	u_int32_t time_stop;		/* 0 < time_end < 24*60*60-1 = 86399 */
-	u_int8_t kerneltime;			/* ignore skb time (and use kerneltime) or not. */
-	time_t date_start;
-	time_t date_stop;
+	unsigned int  days_match;   /* 1 bit per day. -SMTWTFS            */
+	unsigned int  time_start;   /* 0 < time_start   < 86399           */
+	unsigned int   time_stop;    /* 0:0 < time_start < 23:59          */
+	int  kerneltime;   /* ignore skb time (and use kerneltime) or not.*/
+        time_t date_start;
+        time_t date_stop;
 };
